02 Lesson 16 - Fix bags delete && update posts~categorys~tags

+++ 02-baseLaravel/app/Http/Controllers/Admin/CategoryController.php
@@ -99,7 +99,16 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
+        //$category = Category::find($id);
+        //$category->delete();
+
+        //Category::destroy($id);
+
         $category = Category::find($id);
+        //dd($category->posts->count());
+        if($category->posts->count()) {
+            return redirect()->route('categories.index')->with('error', 'Ошибка удаления, есть прикрепленные записи');
+        }
         $category->delete();
         return redirect()->route('categories.index')->with('success', 'Категория удалена');
     }
